Writing tests has completed

+++ src/main/java/hexlet/code/controller/api/LabelController.java
@@ -3,11 +3,14 @@ package hexlet.code.controller.api;
 import hexlet.code.dto.label.LabelCreateDTO;
 import hexlet.code.dto.label.LabelDTO;
 import hexlet.code.dto.label.LabelUpdateDTO;
+import hexlet.code.mapper.ReferenceMapper;
+import hexlet.code.model.Label;
 import hexlet.code.service.LabelService;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -28,33 +31,55 @@ public class LabelController {
     @Autowired
     private LabelService labelService;
 
+    @Autowired
+    private ReferenceMapper referenceMapper;
+
     @GetMapping(path = "")
     @ResponseStatus(HttpStatus.OK)
+    // просматривать метки могут только аутентифицированные пользователи
+    @PreAuthorize("isAuthenticated()")
     public ResponseEntity<List<LabelDTO>> index() {
-        return ResponseEntity.ok(labelService.getAllLabels().getBody());
+        var labels = labelService.getAllLabels();
+
+        return ResponseEntity.ok()
+                .header("X-Total-Count", String.valueOf(labels.size()))
+                .body(labels);
     }
 
     @GetMapping(path = "/{id}")
     @ResponseStatus(HttpStatus.OK)
+    // просматривать метку могут только аутентифицированные пользователи
+    @PreAuthorize("isAuthenticated()")
     public LabelDTO show(@PathVariable long id) {
         return labelService.getLabelById(id);
     }
 
     @PostMapping(path = "")
     @ResponseStatus(HttpStatus.CREATED)
+    // добавлять метку могут только аутентифицированные пользователи
+    @PreAuthorize("isAuthenticated()")
     public LabelDTO create(@Valid @RequestBody LabelCreateDTO labelCreateDTO) {
         return labelService.createLabel(labelCreateDTO);
     }
 
     @PutMapping(path = "/{id}")
     @ResponseStatus(HttpStatus.OK)
+    // обновлять метку могут только аутентифицированные пользователи
+    @PreAuthorize("isAuthenticated()")
     public LabelDTO update(@PathVariable long id, @Valid @RequestBody LabelUpdateDTO labelUpdateDTO) {
         return labelService.updateLabel(id, labelUpdateDTO);
     }
 
     @DeleteMapping(path = "/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
+    // удалять метку могут только аутентифицированные пользователи
+    @PreAuthorize("isAuthenticated()")
     public void delete(@PathVariable long id) {
+        // Если метка связана с задачей, удалить её нельзя.
+        var label = referenceMapper.toEntity(id, Label.class);
+        if (!label.getTasks().isEmpty()) {
+            throw new RuntimeException("Can't delete a label, because it has tasks");
+        }
         labelService.deleteLabel(id);
     }
 }
